Add files via upload
In this project, the behaviour and dynamics of DIPC (Double Inverted Pendulum on a Cart) was simulated, and animated in the 3D fashion. Modeling of the system dynamics was based on the Newton-Euler Mechanics, while the simulation was done by using MATLAB and Simulink. To control and stabilize the pendulum in the unstable upright equilibrium position, a LQR (Linear Quadratic Regulator) controller was used together along with a controller designed based on the Lyapunov theorem. The goal of the controller was to bring the pendulum from the stable equilibrium point up to the upright position, and stabilize it. Finally, animation of the system was shown.